pyproject.toml
[project]
name = "deep-tech-doc-scope"
version = "0.1.0"
description = ""
authors = [
    {name = "Dmitry",email = "kiselevdv1991@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=4.2,<5.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "celery (>=5.5.1,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "whitenoise (>=6.6.0,<7.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"
pytest-cov = "^6.1.1"
flake8 = "^7.2.0"
